refactoring and docu for constraints (#1)

+++ bundles/org.palladiosimulator.spd/src/org/palladiosimulator/spd/adjustments/RelativeAdjustment.java
@@ -9,7 +9,10 @@ package org.palladiosimulator.spd.adjustments;
  * <!-- end-user-doc -->
  *
  * <!-- begin-model-doc -->
- * The RelativeAdjustment denotes that the group should is adjusted relatively to the current number of elements.
+ * The RelativeAdjustment denotes that the group is adjusted relatively to the current number of elements.
+ * The RelativeAdjustment contains two parameters: the percentageGrowthValue and the minAdjustmentValue.
+ * The percentageGrowthValue determines the change (increase/decrease) of the current capacity as a percentage value.
+ * The minAdjustmentValue determines the minimal change of the current capacity.
  * <!-- end-model-doc -->
  *
  * <p>
@@ -27,19 +30,19 @@ package org.palladiosimulator.spd.adjustments;
 public interface RelativeAdjustment extends AdjustmentType {
 	/**
 	 * Returns the value of the '<em><b>Percentage Growth Value</b></em>' attribute.
-	 * The default value is <code>"0.0"</code>.
+	 * The default value is <code>"100"</code>.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * <!-- begin-model-doc -->
 	 * The percantage value of adjustment e.g., a value of 10 denotes that 10% should be added to the existing capacity.
 	 * <!-- end-model-doc -->
 	 * @return the value of the '<em>Percentage Growth Value</em>' attribute.
-	 * @see #setPercentageGrowthValue(double)
+	 * @see #setPercentageGrowthValue(int)
 	 * @see org.palladiosimulator.spd.adjustments.AdjustmentsPackage#getRelativeAdjustment_PercentageGrowthValue()
-	 * @model default="0.0" required="true"
+	 * @model default="100" required="true"
 	 * @generated
 	 */
-	double getPercentageGrowthValue();
+	int getPercentageGrowthValue();
 
 	/**
 	 * Sets the value of the '{@link org.palladiosimulator.spd.adjustments.RelativeAdjustment#getPercentageGrowthValue <em>Percentage Growth Value</em>}' attribute.
@@ -49,7 +52,7 @@ public interface RelativeAdjustment extends AdjustmentType {
 	 * @see #getPercentageGrowthValue()
 	 * @generated
 	 */
-	void setPercentageGrowthValue(double value);
+	void setPercentageGrowthValue(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Min Adjustment Value</b></em>' attribute.
